requirements.txt
python-telegram-bot==21.4
aiohttp==3.9.5
beautifulsoup4==4.12.2
playwright==1.40.0
httpx==0.27.0
